add auth to views

+++ resources/views/permintaan/index.blade.php
@@ -11,9 +11,11 @@
     <h1 class="h3 mb-4 text-gray-800">Permintaan</h1>
     
     <div class="card shadow mb-4">
-      <div class="card-header py-3">
-        <button class="btn btn-primary btn-sm" type="button" onclick="location.assign('{{ route('permintaan.create') }}')">Tambah Baru</button>
-      </div>
+      @if(auth()->user()->role === 'ADMIN')
+        <div class="card-header py-3">
+          <button class="btn btn-primary btn-sm" type="button" onclick="location.assign('{{ route('permintaan.create') }}')">Tambah Baru</button>
+        </div>
+      @endif
       <div class="card-body">
 
         <x-alert-messages />
@@ -27,7 +29,9 @@
                 <th>Permintaan</th>
                 <th>Sisa</th>
                 <th>Kekurangan</th>
-                <th></th>
+                @if(auth()->user()->role === 'ADMIN')
+                  <th></th>
+                @endif
               </tr>
             </thead>
             <tbody>
@@ -38,15 +42,17 @@
                   <td>{{ $p->permintaan }}</td>
                   <td>{{ $p->sisa }}</td>
                   <td>{{ $p->kekurangan }}</td>
-                  <td>
-                    <form action="{{ route('permintaan.destroy', ['permintaan' => $p->id]) }}" method="POST">
-                      @method('DELETE')
-                      @csrf
-                      <a href="{{ route('permintaan.edit', ['permintaan' => $p->id]) }}">Edit</a>
-                      <span class="mx-1 text-black-50">|</span>
-                      <a class="text-danger" href="{{ route('permintaan.destroy', ['permintaan' => $p->id]) }}" onclick="deleteData()">Hapus</a>
-                    </form>
-                  </td>
+                  @if(auth()->user()->role === 'ADMIN')
+                    <td>
+                      <form action="{{ route('permintaan.destroy', ['permintaan' => $p->id]) }}" method="POST">
+                        @method('DELETE')
+                        @csrf
+                        <a href="{{ route('permintaan.edit', ['permintaan' => $p->id]) }}">Edit</a>
+                        <span class="mx-1 text-black-50">|</span>
+                        <a class="text-danger" href="{{ route('permintaan.destroy', ['permintaan' => $p->id]) }}" onclick="deleteData()">Hapus</a>
+                      </form>
+                    </td>
+                  @endif
                 </tr>
               @endforeach
             </tbody>
